requirements.txt
flask
tensorflow
numpy
opencv-python
Pillow
gunicorn
gradio
